+++ service/src/async_fn.rs
@@ -6,8 +6,8 @@ use core::future::Future;
 ///
 /// It is necessary in the the HRTB bounds for async fn's with reference parameters because it
 /// allows the output future to be bound to the parameter lifetime.
-///     `F: for<'a> AsyncClosure<(&'a u8,) Output=u8>`
-pub trait AsyncClosure<Arg> {
+///     `F: for<'a> AsyncFn<(&'a u8,) Output=u8>`
+pub trait AsyncFn<Arg> {
     type Output;
     type Future: Future<Output = Self::Output>;
 
@@ -16,7 +16,7 @@ pub trait AsyncClosure<Arg> {
 
 macro_rules! async_closure_impl {
     ($($arg: ident),*) => {
-        impl<Func, Fut, $($arg,)*> AsyncClosure<($($arg,)*)> for Func
+        impl<Func, Fut, $($arg,)*> AsyncFn<($($arg,)*)> for Func
         where
             Func: Fn($($arg),*) -> Fut,
             Fut: Future,