+++ service/src/middleware/mod.rs
@@ -2,8 +2,10 @@
 //!
 //! [ServiceExt::enclosed]: crate::service::ServiceExt::enclosed
 
+mod async_fn;
 mod group;
 mod unchecked_ready;
 
+pub use async_fn::AsyncFn;
 pub use group::Group;
 pub use unchecked_ready::UncheckedReady;