+++ service/src/pipeline/marker.rs
@@ -7,5 +7,4 @@ pub struct MapErr;
 pub struct BuildAndThen;
 pub struct AndThen;
 pub struct BuildEnclosed;
-pub struct BuildEnclosedFn;
-pub struct EnclosedFn;
+pub struct AsyncFn;