+++ server/CHANGES.md
@@ -1,4 +1,7 @@
-# unreleased
+# unreleased 0.5.0
+
+## Change
+- update `xitca-service` to `0.3.0`
 
 # 0.4.0
 ## Change